platforms/gke/base/use-cases/inference-ref-arch/terraform/comfyui/src/custom-nodes/gemini/requirements.txt
# Copyright 2025 Google LLC
#
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#
#      http://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.

# This is a preview version of gemini custom node
google-api-core==2.24.2
google-genai==1.20.0
google-generativeai==0.8.5
opencv-python-headless==4.11.0.86
